Renamed index.ts to getIndex.ts
The filename index gets set as main entrypoint by npm
