+++ src/getIndex.ts
